+++ bank-feeds/RELEASES.md
@@ -657,3 +657,13 @@ Based on:
 - [typescript v2.1.0] bank-feeds
 ### Releases
 - [NPM v2.1.0] https://www.npmjs.com/package/@codat/bank-feeds/v/2.1.0 - bank-feeds
+
+## 2023-09-21 00:05:24
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.88.1 (2.122.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.1.1] bank-feeds
+### Releases
+- [NPM v2.1.1] https://www.npmjs.com/package/@codat/bank-feeds/v/2.1.1 - bank-feeds